+++ flutter_inappwebview_forge_ios/ios/flutter_inappwebview_forge_ios/Sources/flutter_inappwebview_forge_ios/InAppWebView/WebMessage/WebMessageListener.swift
@@ -128,11 +128,16 @@ public class WebMessageListener: FlutterMethodCallDelegate {
         guard let map = map else {
             return nil
         }
+        guard let id = map["id"] as? String,
+              let jsObjectName = map["jsObjectName"] as? String,
+              let originRules = map["allowedOriginRules"] as? [String] else {
+            return nil
+        }
         return WebMessageListener(
             plugin: plugin,
-            id: map["id"] as! String,
-            jsObjectName: map["jsObjectName"] as! String,
-            allowedOriginRules: Set(map["allowedOriginRules"] as! [String])
+            id: id,
+            jsObjectName: jsObjectName,
+            allowedOriginRules: Set(originRules)
         )
     }